Rename package from lib->common
