+++ test-common/src/main/java/org/triplea/test/common/CustomMatcher.java
@@ -1,4 +1,4 @@
-package org.triplea.test.lib;
+package org.triplea.test.common;
 
 import java.util.function.Function;
